services/user-service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
python-consul==1.1.0
pyjwt==2.8.0
